+++ lib/src/models/customer/create_customer_response_model.dart
@@ -1,3 +1,6 @@
+import 'package:collection/collection.dart';
+import 'package:mx_merchant/mx_merchant.dart';
+
 class MxCreateCustomerResponseModel {
   final int id;
   final String? created;
@@ -24,7 +27,7 @@ class MxCreateCustomerResponseModel {
   final String? city;
   final String? state;
   final String? zip;
-  final String? customerType;
+  final MxCustomerTypeEnum customerType;
   final bool spendProfileOverride;
   final bool isTaxExempt;
   final bool hasPayments;
@@ -32,6 +35,11 @@ class MxCreateCustomerResponseModel {
   final bool hasInvoices;
   final String? displayColor;
   final bool activeStatus;
+  final String? email;
+  final String? mobile;
+  final String? phone;
+  final List<MxCustomerCustomFieldModel>? customFields;
+  String? selectedSpendProfile;
 
   MxCreateCustomerResponseModel({
     this.id = 0,
@@ -59,7 +67,7 @@ class MxCreateCustomerResponseModel {
     this.city,
     this.state,
     this.zip,
-    this.customerType,
+    this.customerType = .person,
     this.spendProfileOverride = true,
     this.isTaxExempt = true,
     this.hasPayments = true,
@@ -67,6 +75,11 @@ class MxCreateCustomerResponseModel {
     this.hasInvoices = true,
     this.displayColor,
     this.activeStatus = true,
+    this.email,
+    this.mobile,
+    this.phone,
+    this.customFields,
+    this.selectedSpendProfile,
   });
 
   factory MxCreateCustomerResponseModel.fromJson(Map<String, dynamic> json) {
@@ -96,7 +109,7 @@ class MxCreateCustomerResponseModel {
       city: json['city'],
       state: json['state'],
       zip: json['zip'],
-      customerType: json['customerType'],
+      customerType: MxCustomerTypeEnum.values.firstWhereOrNull((e) => e.name == json['customerType']) ?? MxCustomerTypeEnum.person,
       spendProfileOverride: json['spendProfileOverride'] ?? true,
       isTaxExempt: json['isTaxExempt'] ?? true,
       hasPayments: json['hasPayments'] ?? true,
@@ -104,6 +117,11 @@ class MxCreateCustomerResponseModel {
       hasInvoices: json['hasInvoices'] ?? true,
       displayColor: json['displayColor'],
       activeStatus: json['activeStatus'] ?? true,
+      email: json['email'],
+      mobile: json['mobile'],
+      phone: json['phone'],
+      customFields: json['customFields'] != null ? MxCustomerCustomFieldModel.fromJsonArray(json['customFields']) : null,
+      selectedSpendProfile: json['selectedSpendProfile'],
     );
   }
 
@@ -134,7 +152,7 @@ class MxCreateCustomerResponseModel {
       'city': city,
       'state': state,
       'zip': zip,
-      'customerType': customerType,
+      'customerType': customerType.name,
       'spendProfileOverride': spendProfileOverride,
       'isTaxExempt': isTaxExempt,
       'hasPayments': hasPayments,
@@ -142,6 +160,97 @@ class MxCreateCustomerResponseModel {
       'hasInvoices': hasInvoices,
       'displayColor': displayColor,
       'activeStatus': activeStatus,
+      'email': email,
+      'mobile': mobile,
+      'phone': phone,
+      'customFields': customFields?.map((e) => e.toJson()).toList(),
+      'selectedSpendProfile': selectedSpendProfile,
     }..removeWhere((_, v) => v == null);
+  }
+
+  MxCreateCustomerResponseModel copyWith({
+    int? id,
+    String? created,
+    String? name,
+    String? firstName,
+    String? lastName,
+    String? number,
+    String? taxId,
+    int? visits,
+    int? transactionCount,
+    String? spend,
+    String? sales,
+    String? refunds,
+    String? lastActivity,
+    bool? allowPromotions,
+    String? loyaltyStatus,
+    String? loyaltyEnrolled,
+    bool? isVip,
+    String? spendProfile,
+    String? spendProfileIcon,
+    String? addressName,
+    String? address1,
+    String? address2,
+    String? city,
+    String? state,
+    String? zip,
+    MxCustomerTypeEnum? customerType,
+    bool? spendProfileOverride,
+    bool? isTaxExempt,
+    bool? hasPayments,
+    bool? hasContracts,
+    bool? hasInvoices,
+    String? displayColor,
+    bool? activeStatus,
+    String? email,
+    String? mobile,
+    String? phone,
+    List<MxCustomerCustomFieldModel>? customFields,
+    String? selectedSpendProfile,
+  }) {
+    return MxCreateCustomerResponseModel(
+      id: id ?? this.id,
+      created: created ?? this.created,
+      name: name ?? this.name,
+      firstName: firstName ?? this.firstName,
+      lastName: lastName ?? this.lastName,
+      number: number ?? this.number,
+      taxId: taxId ?? this.taxId,
+      visits: visits ?? this.visits,
+      transactionCount: transactionCount ?? this.transactionCount,
+      spend: spend ?? this.spend,
+      sales: sales ?? this.sales,
+      refunds: refunds ?? this.refunds,
+      lastActivity: lastActivity ?? this.lastActivity,
+      allowPromotions: allowPromotions ?? this.allowPromotions,
+      loyaltyStatus: loyaltyStatus ?? this.loyaltyStatus,
+      loyaltyEnrolled: loyaltyEnrolled ?? this.loyaltyEnrolled,
+      isVip: isVip ?? this.isVip,
+      spendProfile: spendProfile ?? this.spendProfile,
+      spendProfileIcon: spendProfileIcon ?? this.spendProfileIcon,
+      addressName: addressName ?? this.addressName,
+      address1: address1 ?? this.address1,
+      address2: address2 ?? this.address2,
+      city: city ?? this.city,
+      state: state ?? this.state,
+      zip: zip ?? this.zip,
+      customerType: customerType ?? this.customerType,
+      spendProfileOverride: spendProfileOverride ?? this.spendProfileOverride,
+      isTaxExempt: isTaxExempt ?? this.isTaxExempt,
+      hasPayments: hasPayments ?? this.hasPayments,
+      hasContracts: hasContracts ?? this.hasContracts,
+      hasInvoices: hasInvoices ?? this.hasInvoices,
+      displayColor: displayColor ?? this.displayColor,
+      activeStatus: activeStatus ?? this.activeStatus,
+      email: email ?? this.email,
+      mobile: mobile ?? this.mobile,
+      phone: phone ?? this.phone,
+      customFields: customFields ?? this.customFields,
+      selectedSpendProfile: selectedSpendProfile ?? this.selectedSpendProfile,
+    );
+  }
+
+  MxCreateCustomerRequestModel toRequestModel() {
+    return MxCreateCustomerRequestModel.fromJson(toJson());
   }
 }